******************************************************************************** ********************************************************************************
# Changes

## Info

**Date & Time**: 03-SEP-2023

**Author**:      Pablo Gomez Martino

## Work Items

* Task 1: ACC LIS3D first approach [COMPLETE]
* Task 1: ACC LIS2D first approach [COMPLETE]
* Task 1: MG LIS2MDL first approach [COMPLETE]
* Task 1: PR LPS22HH first approach [COMPLETE]

## Details
* Task 1:
  * None

## Notes:

* None
********************************************************************************
********************************************************************************

+++ drivers/i2c_low_level_driver/i2c_low_level_driver.h
@@ -10,7 +10,7 @@
 
 #include "sl_i2cspm.h"
 
-I2C_TransferReturn_TypeDef i2c_slave_read  (I2C_TypeDef *i2c, uint8_t addr, uint8_t command, uint8_t *val, uint8_t len);
-I2C_TransferReturn_TypeDef i2c_slave_write (I2C_TypeDef *i2c, uint8_t addr, uint8_t command, uint8_t *val, uint8_t len);
+I2C_TransferReturn_TypeDef i2c_slave_read  (I2C_TypeDef *i2c, uint8_t addr, uint16_t command, uint8_t *val, uint8_t len);
+I2C_TransferReturn_TypeDef i2c_slave_write (I2C_TypeDef *i2c, uint8_t addr, uint16_t command, uint8_t *val, uint8_t len);
 
 #endif /* DRIVERS_I2C_LOW_LEVEL_DRIVER_I2C_LOW_LEVEL_DRIVER_H_ */
